Added custom exceptions and logging with requestId

+++ src/main/java/spring/ru/springtest/services/AuthorService.java
@@ -1,9 +1,11 @@
 package spring.ru.springtest.services;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import spring.ru.springtest.dto.Author;
+import spring.ru.springtest.exceptions.AuthorNotFoundException;
 import spring.ru.springtest.mapper.AuthorMapper;
 import spring.ru.springtest.models.AuthorModel;
 import spring.ru.springtest.repositories.AuthorRepository;
@@ -11,46 +13,57 @@ import spring.ru.springtest.repositories.AuthorRepository;
 import java.util.UUID;
 
 @Service
-@Transactional
+@RequiredArgsConstructor
+@Slf4j
 public class AuthorService {
 
     private final AuthorRepository authorRepository;
     private final AuthorMapper authorMapper;
 
-    @Autowired
-    public AuthorService(AuthorRepository authorRepository, AuthorMapper authorMapper) {
-        this.authorRepository = authorRepository;
-        this.authorMapper = authorMapper;
-    }
-
     @Transactional(readOnly = true)
-    public AuthorModel findById(UUID id) {
-        return authorRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Author not found"));
+    public Author findById(UUID id) {
+        log.debug("Finding author by id: {}", id);
+        AuthorModel authorModel = authorRepository.findById(id)
+                .orElseThrow(() -> {
+                    log.error("Author not found with id {}", id);
+                    return new AuthorNotFoundException("Author with id " + id + " not found");
+                });
+
+        return authorMapper.toDto(authorModel);
     }
 
-    public void save(AuthorModel author) {
-        if (author.getBooks() != null) {
-            author.getBooks().forEach(book -> book.setAuthor(author));
-        }
+    @Transactional
+    public void save(Author authorDto) {
+        log.debug("Saving author: {}", authorDto);
+        AuthorModel authorModel = authorMapper.toEntity(authorDto);
 
-        authorRepository.save(author);
+        authorRepository.save(authorModel);
+        log.info("Saved author with id {}", authorModel.getId());
     }
 
+    @Transactional
     public void update(UUID id, Author dto) {
+        log.debug("Update author with id: {}", id);
         AuthorModel existingAuthor = authorRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Author not found"));
+                .orElseThrow(() -> {
+                    log.error("Update failed: author not found with id {}", id);
+                    return new AuthorNotFoundException("Author with id " + id + " not found");
+                });
 
         authorMapper.updateEntityFromDto(dto, existingAuthor);
-
-        if (existingAuthor.getBooks() != null) {
-            existingAuthor.getBooks().forEach(book -> book.setAuthor(existingAuthor));
-        }
-
         authorRepository.save(existingAuthor);
+        log.info("Updated author with id {}", id);
     }
 
+    @Transactional
     public void delete(UUID id) {
+
+        log.debug("Delete author with id {}", id);
+        if (!authorRepository.existsById(id)) {
+            log.warn("Attempt to delete non-existent author with id: {}", id);
+            throw new AuthorNotFoundException("Author with id " + id + " not found");
+        }
         authorRepository.deleteById(id);
+        log.info("Deleted author with id {}", id);
     }
 }
